requirements.txt
Flask==3.0.3
pytesseract==0.3.10
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pdf2image==1.17.0
spacy==3.7.5
networkx==3.3
google-cloud-vision==3.7.2